add comments, move imports
