+++ Homework 3/api_gateway/auth/exceptions.py
