+++ Homework 3/api_gateway/courses/service.py
@@ -1,11 +1,12 @@
 from typing import Optional
-from api_gateway.courses.schemas import CourseShort, CourseFull, Lesson
 from protos.course_pb2_grpc import CourseServiceStub
 from protos.course_pb2 import CourseFilterRequest, CourseRequest, CourseListResponse, CourseFullResponse
+from api_gateway.courses.schemas import CourseShort, CourseFull, Lesson
 from api_gateway.courses.exceptions import CourseDoesntExist
 
 
 class CourseService:
+    """Сервис для курсов через gRPC"""
     def __init__(self, stub: CourseServiceStub):
         self.stub = stub
 
@@ -14,6 +15,7 @@ class CourseService:
             implementer: Optional[str] = None,
             role: Optional[str] = None,
             search: Optional[str] = None) -> list[CourseShort]:
+        """Возвращает курсы по переданным фильтрам"""
         response: CourseListResponse = await self.stub.get_courses(
             CourseFilterRequest(implementer=implementer, role=role, search=search)
         )
@@ -23,6 +25,7 @@ class CourseService:
         ]
 
     async def get_course_info(self, course_id: int) -> CourseFull:
+        """Возвращает подробную информацию о курсе или вызывает ошибку"""
         response: CourseFullResponse = await self.stub.get_course_info(CourseRequest(course_id=course_id))
         if response.id == 0:
             raise CourseDoesntExist
@@ -32,7 +35,7 @@ class CourseService:
             description=response.description,
             is_prerecorded_course=response.is_prerecorded_course,
             implementer=response.implementer,
-            roles=[role for role in response.roles],
+            roles=list(response.roles),
             lessons=[
                 Lesson(
                     number=lesson.number, day_of_week=lesson.day_of_week, string_day=lesson.string_day,
