+++ Homework 3/api_gateway/settings.py
@@ -4,6 +4,7 @@ from dotenv import load_dotenv
 
 
 class Settings(BaseSettings):
+    """Конфигурация для микросервиса"""
     course_grpc_server_address: str = Field(alias="COURSE_ADDRESS")
     auth_grpc_server_address: str = Field(alias="AUTH_ADDRESS")
     time_checking_grpc_server_address: str = Field(alias="TIME_CHECKING_ADDRESS")
