+++ Homework 3/auth_microservice/service/utils.py
@@ -2,4 +2,5 @@ import hashlib
 
 
 def hash_password(password: str) -> str:
+    """Выполняет хеширование пароля"""
     return hashlib.sha256(password.encode("utf-8")).hexdigest()
