+++ Homework 3/auth_microservice/settings.py
@@ -4,6 +4,7 @@ from dotenv import load_dotenv
 
 
 class Settings(BaseSettings):
+    """Конфигурация для микросервиса"""
     auth_grpc_server_address: str = Field(alias="AUTH_ADDRESS")
     db_filepath: str = Field(alias="AUTH_DB_FILEPATH")
     mode: str = Field(alias="MODE")
